Moves the permission labels helpers to logic.auth
