+++ ckanext/knowledgehub/logic/auth/__init__.py
@@ -1,0 +1,4 @@
+from ckanext.knowledgehub.logic.auth.permissions import (
+    get_permission_labels,
+    get_user_permission_labels,
+)
